requirements.txt
streamlit>=1.29.0
pymupdf4llm>=0.0.10
PyMuPDF>=1.23.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10